Refactor code - CommentData, Listings, and Submissions derive Default   - Easier to init structs with a few values - SubmissionData and Submission have ::new() functions for defaults - Web configuration functions and structs in dedicated module/file   - Errors, functions, etc. commonly used - api_translation takes advantage of defaults in some functions - Endpoint info in dedicated module/file   - Add default "/" route, at least for now - Tested locally for endpoint `/hot.json`
